Add "Linear" prefix to UndoManagerImpl/Factory and move all linear-related classes into separate package
