+++ undofx/src/main/java/org/fxmisc/undo/LinearUndoManagerFactory.java
@@ -6,13 +6,13 @@ import java.util.function.Consumer;
 import java.util.function.Function;
 
 import org.fxmisc.undo.impl.ChangeQueue;
-import org.fxmisc.undo.impl.FixedSizeLinearChangeQueue;
-import org.fxmisc.undo.impl.UndoManagerImpl;
-import org.fxmisc.undo.impl.UnlimitedLinearChangeQueue;
-import org.fxmisc.undo.impl.ZeroSizeLinearChangeQueue;
+import org.fxmisc.undo.impl.linear.FixedSizeLinearChangeQueue;
+import org.fxmisc.undo.impl.linear.LinearUndoManagerImpl;
+import org.fxmisc.undo.impl.linear.UnlimitedLinearChangeQueue;
+import org.fxmisc.undo.impl.linear.ZeroSizeLinearChangeQueue;
 import org.reactfx.EventStream;
 
-public interface UndoManagerFactory {
+public interface LinearUndoManagerFactory {
 
     <C> UndoManager create(
             EventStream<C> changeStream,
@@ -31,7 +31,7 @@ public interface UndoManagerFactory {
             Consumer<C> apply) {
         ChangeQueue<C> queue = new UnlimitedLinearChangeQueue<C>();
         BiFunction<C, C, Optional<C>> merge = (c1, c2) -> Optional.empty();
-        return new UndoManagerImpl<>(queue, invert, apply, merge, changeStream);
+        return new LinearUndoManagerImpl<>(queue, invert, apply, merge, changeStream);
     }
 
     public static <C> UndoManager unlimitedHistoryUndoManager(
@@ -40,11 +40,11 @@ public interface UndoManagerFactory {
             Consumer<C> apply,
             BiFunction<C, C, Optional<C>> merge) {
         ChangeQueue<C> queue = new UnlimitedLinearChangeQueue<C>();
-        return new UndoManagerImpl<>(queue, invert, apply, merge, changeStream);
+        return new LinearUndoManagerImpl<>(queue, invert, apply, merge, changeStream);
     }
 
-    public static UndoManagerFactory unlimitedHistoryFactory() {
-        return new UndoManagerFactory() {
+    public static LinearUndoManagerFactory unlimitedHistoryFactory() {
+        return new LinearUndoManagerFactory() {
             @Override
             public <C> UndoManager create(
                     EventStream<C> changeStream,
@@ -71,7 +71,7 @@ public interface UndoManagerFactory {
             int capacity) {
         ChangeQueue<C> queue = new FixedSizeLinearChangeQueue<C>(capacity);
         BiFunction<C, C, Optional<C>> merge = (c1, c2) -> Optional.empty();
-        return new UndoManagerImpl<>(queue, invert, apply, merge, changeStream);
+        return new LinearUndoManagerImpl<>(queue, invert, apply, merge, changeStream);
     }
 
     public static <C> UndoManager fixedSizeHistoryUndoManager(
@@ -81,11 +81,11 @@ public interface UndoManagerFactory {
             BiFunction<C, C, Optional<C>> merge,
             int capacity) {
         ChangeQueue<C> queue = new FixedSizeLinearChangeQueue<C>(capacity);
-        return new UndoManagerImpl<>(queue, invert, apply, merge, changeStream);
+        return new LinearUndoManagerImpl<>(queue, invert, apply, merge, changeStream);
     }
 
-    public static UndoManagerFactory fixedSizeHistoryFactory(int capacity) {
-        return new UndoManagerFactory() {
+    public static LinearUndoManagerFactory fixedSizeHistoryFactory(int capacity) {
+        return new LinearUndoManagerFactory() {
             @Override
             public <C> UndoManager create(
                     EventStream<C> changeStream,
@@ -107,11 +107,11 @@ public interface UndoManagerFactory {
 
     public static <C> UndoManager zeroHistoryUndoManager(EventStream<C> changeStream) {
         ChangeQueue<C> queue = new ZeroSizeLinearChangeQueue<>();
-        return new UndoManagerImpl<>(queue, c -> c, c -> {}, (c1, c2) -> Optional.empty(), changeStream);
+        return new LinearUndoManagerImpl<>(queue, c -> c, c -> {}, (c1, c2) -> Optional.empty(), changeStream);
     }
 
-    public static UndoManagerFactory zeroHistoryFactory() {
-        return new UndoManagerFactory() {
+    public static LinearUndoManagerFactory zeroHistoryFactory() {
+        return new LinearUndoManagerFactory() {
             @Override
             public <C> UndoManager create(
                     EventStream<C> changeStream,
